ResourceModel added in models.py re-factored created_at and updated_at from models.py
